+++ aim/web/ui/src/components/ImagesList/ImageBox.tsx
@@ -1,8 +1,9 @@
-import React, { memo, useEffect } from 'react';
+import React, { memo, MouseEvent, useEffect, useRef } from 'react';
+import { isEqual } from 'lodash-es';
 
 import { Skeleton } from '@material-ui/lab';
 
-import { imageFixedHeight } from 'config/imagesConfigs/imagesConfig';
+import { batchCollectDelay } from 'config/imagesConfigs/imagesConfig';
 
 const ImageBox = ({
   index,
@@ -10,11 +11,16 @@ const ImageBox = ({
   data,
   imagesBlobs,
   addUriToList,
+  imageHeight,
+  focusedState,
+  syncHoverState,
 }: any): React.FunctionComponentElement<React.ReactNode> => {
   const { format, blob_uri } = data;
 
   useEffect(() => {
-    let timeoutID = setTimeout(() => addUriToList(blob_uri), 900);
+    let timeoutID = setTimeout(() => {
+      addUriToList(blob_uri);
+    }, batchCollectDelay);
 
     return () => {
       clearTimeout(timeoutID);
@@ -22,9 +28,50 @@ const ImageBox = ({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
+  function safeSyncHoverState(args: any): void {
+    if (typeof syncHoverState === 'function') {
+      syncHoverState(args);
+    }
+  }
+
+  function onClick(e: MouseEvent<HTMLDivElement>): void {
+    // TODO need to add focused image logic
+    // if (e?.currentTarget) {
+    //   e.stopPropagation();
+    //   const clientRect = e.currentTarget.getBoundingClientRect();
+    //   safeSyncHoverState({
+    //     activePoint: { clientRect, key: data.key, seqKey: data.seqKey },
+    //     focusedStateActive: true,
+    //   });
+    // }
+  }
+
+  function onMouseEnter(e: MouseEvent<HTMLDivElement>): void {
+    if (e?.currentTarget && !focusedState?.active) {
+      const clientRect = e.currentTarget.getBoundingClientRect();
+      safeSyncHoverState({
+        activePoint: { clientRect, key: data.key, seqKey: data.seqKey },
+      });
+    }
+  }
+
+  function onMouseLeave(e: MouseEvent<HTMLDivElement>): void {
+    if (!focusedState?.active) {
+      safeSyncHoverState({ activePoint: null });
+    }
+  }
+
   return (
     <div key={index} className='ImagesSet__container__imagesBox__imageBox'>
-      <div style={style}>
+      <div
+        style={style}
+        className={
+          focusedState?.active && focusedState.key === data.key ? 'active' : ''
+        }
+        onMouseEnter={onMouseEnter}
+        onMouseLeave={onMouseLeave}
+        onClick={onClick}
+      >
         {imagesBlobs?.[blob_uri] ? (
           <img
             src={`data:image/${format};base64, ${imagesBlobs?.[blob_uri]}`}
@@ -33,7 +80,7 @@ const ImageBox = ({
         ) : (
           <Skeleton
             variant='rect'
-            height={imageFixedHeight - 10}
+            height={imageHeight - 10}
             width={style.width - 10}
           />
         )}