+++ aim/web/ui/src/components/ImagesList/ImagesList.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import { VariableSizeList as List } from 'react-window';
 
-import { imageFixedHeight } from 'config/imagesConfigs/imagesConfig';
-
 import ImageBox from './ImageBox';
 
 function ImagesList({
@@ -11,13 +9,16 @@ function ImagesList({
   onScroll,
   imageSetWrapperWidth,
   addUriToList,
+  imageHeight,
+  focusedState,
+  syncHoverState,
 }: any): React.FunctionComponentElement<React.ReactNode> {
   return (
     <List
-      height={imageFixedHeight}
+      height={imageHeight}
       itemCount={data.length}
       itemSize={(index: number) =>
-        (imageFixedHeight / data[index].height) * data[index].width
+        (imageHeight / data[index].height) * data[index].width
       }
       layout='horizontal'
       width={imageSetWrapperWidth || 0}
@@ -31,6 +32,9 @@ function ImagesList({
           data={data[index]}
           imagesBlobs={imagesBlobs}
           addUriToList={addUriToList}
+          imageHeight={imageHeight}
+          focusedState={focusedState}
+          syncHoverState={syncHoverState}
         />
       )}
     </List>