+++ aim/web/ui/src/components/ImagesPanel/ImagesPanel.d.ts
@@ -1,7 +1,15 @@
 import React from 'react';
 
+import { ResizeModeEnum } from 'config/enums/tableEnums';
+
+import {
+  IFocusedState,
+  ITooltipContent,
+} from 'types/services/models/metrics/metricsAppModel';
+
 export interface IImagesPanelProps {
   imagesData: any;
+  orderedMap: { [key: string]: any };
   imagesBlobs: { [key: string]: value };
   recordSlice: number[];
   indexSlice: number[];
@@ -18,4 +26,9 @@ export interface IImagesPanelProps {
   panelResizing: boolean;
   imageWrapperOffsetHeight: number;
   imageWrapperOffsetWidth: number;
+  controls?: React.ReactNode;
+  resizeMode: ResizeModeEnum;
+  tooltip: ITooltipContent;
+  focusedState: IFocusedState;
+  onActivePointChange?: (activePoint: any, focusedStateActive: boolean) => void;
 }