+++ aim/web/ui/src/components/ImagesSet/ImagesSet.d.ts
@@ -1,6 +1,8 @@
+import { IFocusedState } from 'types/services/models/metrics/metricsAppModel';
+
 export interface IImageSetProps {
   data: any;
-  title: string;
+  orderedMap: { [key: string]: any };
   imagesBlobs: object;
   onScroll: () => void;
   addUriToList: (blobUrl: string) => void;
@@ -8,4 +10,7 @@ export interface IImageSetProps {
   imagesSetKey: number;
   imageSetWrapperHeight?: number;
   imageSetWrapperWidth?: number;
+  imageHeight: number;
+  syncHoverState?: (args: any) => void;
+  focusedState: IFocusedState;
 }